third_party/WebKit/LayoutTests/virtual/threaded/fast/compositorworker/basic-plumbing-main-to-worker.html

<!DOCTYPE html>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script>
if (window.internals)
  internals.settings.setCompositorWorkerEnabled(true);

var test = async_test('Tests that a change from the main thread is received on the compositor thread');
var worker;

test.step(function(){
  worker = new CompositorWorker('resources/basic-plumbing-main-to-worker.js');
  worker.onmessage = test.step_func(function(e){
    assert_equals(e.data.value, 42);
    test.done();
  });
  worker.postMessage({value: 42});
});
</script>
